Makfile error on Linux

+++ source/class/give_socket_name.cpp
@@ -6,7 +6,7 @@
 /*   By: tda-silv <tda-silv@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/01 15:03:33 by tda-silv          #+#    #+#             */
-/*   Updated: 2023/08/04 14:27:44 by tda-silv         ###   ########.fr       */
+/*   Updated: 2023/08/07 09:24:04 by tda-silv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ int	Server::give_socket_name(struct sockaddr_in *adress, const int PORT, int ser
 	memset( (char *) adress, 0, sizeof(*adress) );
 	adress->sin_family = AF_INET;
 	adress->sin_addr.s_addr = INADDR_ANY; 
-	adress->sin_port = htons(PORT);
+	adress->sin_port = htons(static_cast<uint16_t>(PORT) );
 
 	if (bind(server_fd, (struct sockaddr *) adress, sizeof(*adress) ) < 0) 
 	{
